Set and prepare interval for exp data

Gathers data needed for the plotting of drug administration periods and transforms them to a consistent structure.

Usage

set_exp_intervals(data_list, mapping = default_drug_admin(), subjid_var)

Arguments

data_list

A named list of data frames that provides a subject level dataset and further datasets with event specific data. Usually obtained from module manager.

mapping

A list of named character strings that describes which variables to use to display drug administration events.

subjid_var

[character(1)]

Character name of the unique subject identifier column in all datasets (default is USUBJID). Must be a single value.

Value

A data frame including the following columns:

subject_id: A unique subject identifier column. Its name is specified by subjid_var.

start_exp, end_exp: Start and end dates of the event.

detail_var: Information about drug name and dosage.

set: Name of the dataset the data origins from.

set_id: Row ID's of the related dataset.

dose_var: Indicates whether the dose of the treatment has increased, decreased or stayed the same compared to the last dose at the same subject.

group: Label for the event types.

Details

The mapping list must contain the following elements:

name: Character name of the dataset that holds drug administration data (e.g. ex domain), as it is called in the datalist that is provided to the modulemanager.

start_var: Character name of the variable that contains the start dates (e.g. exposure start dates) which must be present in the dataset mentioned in the name element.

end_var: Character name of the variable that contains the end dates (e.g. exposure end dates) which must be present in the dataset mentioned in the name element.

detail_var: Character name of the variable that contains the treatment information. Must exist in the dataset mentioned in the name element.

label: Free-text character label for the drug administration event.

dose_var: Character name of the variable that contains the dosis level information. Must exist in the dataset mentioned in the name element.

dose_unit_var: Character name of the variable that contains the dosis unit. Must exist in the dataset mentioned in the name element.